Create bitwiseManipulator_P5.py
The user enters an operation they want to perform, how many numbers they want to input and finally are prompted for what each of those integers are (in hexadecimal form). The program will perform and display a bitwise operation with all the inputs. Then, the program will convert the numbers to binary to perform and display them. Users can continue to do more operations or enter 'q' to quit.